requirements.txt
pandas>=2.0.0
scipy>=1.10.0
numpy>=1.24.0
PyQt6>=6.5.0
pyqtgraph>=0.13.0
PyYAML>=6.0

